Add JSON output to script and change issue template
This change adds a new `--json-output` flag to the script, which
outputs the vulnerabilities found as a JSON string. It also
adapts the GH Action to use it, instead of manually parsing the
original output.

It also changes the template for the created issues, which now
contains the dependency name in its title, and a description
that only mentions the relevant vulnerability (instead of the
full output of the script).